Put one symbol per file
