+++ src/injector/injection-class.interface.ts
@@ -1,0 +1,3 @@
+export interface InjectionClass<T> extends Function {
+  new(...args: any[]): T
+}
