+++ src/injector/injection-selector.type.ts
@@ -1,0 +1,4 @@
+import { InjectionClass } from './injection-class.interface'
+import { InjectionToken } from './injection-token.class'
+
+export type InjectionSelector<T> = InjectionToken<T> | InjectionClass<T>
